Renaming ButtonV2 to ButtonNext because of autotools sanity
