+++ packages/wix-ui-core/src/components/Button-next/button-next.spec.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import { ReactDOMTestContainer } from "../../../test/dom-test-container";
-import { buttonDriverFactory } from "./button-v2.driver";
-import { buttonPrivateDriverFactory } from "./button-v2.driver.private";
-import { ButtonV2 } from "./";
+import { buttonDriverFactory } from "./button-next.driver";
+import { buttonPrivateDriverFactory } from "./button-next.driver.private";
+import { ButtonNext } from "./";
 
-describe("ButtonV2", () => {
+describe("ButtonNext", () => {
   const createDriver = new ReactDOMTestContainer()
     .unmountAfterEachTest()
     .createLegacyRenderer(buttonDriverFactory);
@@ -16,7 +16,7 @@ describe("ButtonV2", () => {
   describe("onClick prop", () => {
     it("should be called on click", () => {
       const onClick = jest.fn();
-      const driver = createDriver(<ButtonV2 onClick={onClick} />);
+      const driver = createDriver(<ButtonNext onClick={onClick} />);
       driver.click();
       expect(onClick).toBeCalled();
     });
@@ -26,13 +26,13 @@ describe("ButtonV2", () => {
     const suffix = <div>suffix</div>;
     const prefix = <div>prefix</div>;
     it("should render suffix when given", () => {
-      const driver = createPrivateDriver(<ButtonV2 suffixIcon={suffix} />);
+      const driver = createPrivateDriver(<ButtonNext suffixIcon={suffix} />);
       expect(driver.suffixExists()).toBeTruthy();
       expect(driver.prefixExists()).toBeFalsy();
     });
 
     it("should render prefix when given", () => {
-      const driver = createPrivateDriver(<ButtonV2 prefixIcon={prefix} />);
+      const driver = createPrivateDriver(<ButtonNext prefixIcon={prefix} />);
       expect(driver.prefixExists()).toBeTruthy();
       expect(driver.suffixExists()).toBeFalsy();
     });
