+++ packages/wix-ui-core/src/components/Button-next/index.ts
@@ -1,0 +1,1 @@
+export { ButtonNext } from "./button-next";
